Add files via upload
In this project i make Navbar With help of HTML CSS 
I hope u like it .
A Navbar is a navigation header that is placed at the top of the page which helps to make it more user-friendly so that the navigation through the website becomes easy and the user can directly search for the topic of their interest. The . navbar class is used to create a navigation bar.