Add files via upload
Responsive footer
A footer is an essential component of a web page, often containing important information such as contact details, site map, or social media links. A responsive footer, on the other hand, is designed to adjust its layout accordingly when viewed on different screen sizes, enhancing the user experience on mobile devices.